Improvements to my tmux setup
Using ctrl-a as my prefix simplifies many things. So we use that
instead of the default one and we remove the special bindings to
handle pane-selection (now we just use the default prefix+arrow).
The change in prefix made it necessary to handle the conflicting
"go to beginning of line" in emacs-mode, so a fix was put in
where we basically just send it twice and everything is ok.

Now we are also using a plugin manager and enabled some gnarly
plugins. the following flow of prefix-/ -> search -> ctrl-o or O
is wonderful and enables me to edit files quickly.